Added migrations to allow tests to support Django 1.8+.
