+++ mutant/contrib/numeric/models.py
@@ -12,42 +12,42 @@ class _NumericMeta:
 
 class SmallIntegerFieldDefinition(FieldDefinition):
     class Meta(_NumericMeta):
-        app_label = 'mutant'
+        app_label = 'numeric'
         proxy = True
         defined_field_class = fields.SmallIntegerField
 
 
 class PositiveSmallIntegerFieldDefinition(FieldDefinition):
     class Meta(_NumericMeta):
-        app_label = 'mutant'
+        app_label = 'numeric'
         proxy = True
         defined_field_class = fields.PositiveSmallIntegerField
 
 
 class IntegerFieldDefinition(FieldDefinition):
     class Meta(_NumericMeta):
-        app_label = 'mutant'
+        app_label = 'numeric'
         proxy = True
         defined_field_class = fields.IntegerField
 
 
 class PositiveIntegerFieldDefinition(FieldDefinition):
     class Meta(_NumericMeta):
-        app_label = 'mutant'
+        app_label = 'numeric'
         proxy = True
         defined_field_class = fields.PositiveIntegerField
 
 
 class BigIntegerFieldDefinition(FieldDefinition):
     class Meta(_NumericMeta):
-        app_label = 'mutant'
+        app_label = 'numeric'
         proxy = True
         defined_field_class = fields.BigIntegerField
 
 
 class FloatFieldDefinition(FieldDefinition):
     class Meta(_NumericMeta):
-        app_label = 'mutant'
+        app_label = 'numeric'
         proxy = True
         defined_field_class = fields.FloatField
 
@@ -68,6 +68,6 @@ class DecimalFieldDefinition(FieldDefinition):
     objects = FieldDefinitionManager()
 
     class Meta(_NumericMeta):
-        app_label = 'mutant'
+        app_label = 'numeric'
         defined_field_class = fields.DecimalField
         defined_field_options = ('max_digits', 'decimal_places',)
