+++ setup.cfg
@@ -1,5 +1,6 @@
 [flake8]
 max-line-length = 119
+exclude = */migrations/*
 
 [isort]
 combine_as_imports=true
